Move program sidebar view
Also updated some styles that were conflicting
