--- src/js/views/admin/programs/sidebar/program-details.hbs
@@ -1,2 +1,0 @@
-<textarea class="input-secondary textarea-flex__input js-input" placeholder="{{ @intl.admin.sidebar.programDetailsTemplate.placeholder }}">{{ details }}</textarea>
-<div class="textarea-flex__container input-secondary program-sidebar__details js-spacer">{{ details }} </div>
