--- src/js/views/admin/programs/sidebar/program-name.hbs
@@ -1,5 +1,0 @@
-<div class="pos--relative">
-  <textarea class="input-primary textarea-flex__input js-input{{#if error}} has-error{{/if}}"{{#if isNew}} placeholder="{{ @intl.admin.sidebar.programNameTemplate.placeholder }}"{{/if}}>{{ name }}</textarea>
-  <div class="textarea-flex__container input-primary js-spacer">{{ name }} </div>
-</div>
-{{#if error}}<div class="form-error u-margin--b-4">{{ error }}</div>{{/if}}
