+++ src/js/views/admin/sidebar/program/program-details.hbs
@@ -1,0 +1,2 @@
+<textarea class="input-secondary textarea-flex__input js-input" placeholder="{{ @intl.admin.sidebar.program.programDetailsTemplate.placeholder }}">{{ details }}</textarea>
+<div class="textarea-flex__container input-secondary programs-sidebar__details js-spacer">{{ details }} </div>
