+++ src/js/views/admin/sidebar/program/program-sidebar.hbs
@@ -1,7 +1,7 @@
 <div class="flex-grow">
   <div class="flex">
     <div class="flex-grow">
-      <h3 class="program-sidebar__heading">{{ @intl.admin.sidebar.programSidebarTemplate.headingText }}</h3>
+      <h3 class="programs-sidebar__heading">{{ @intl.admin.sidebar.program.programSidebarTemplate.headingText }}</h3>
     </div>
     <div>
       <button class="button--icon u-margin--r-8 js-menu">{{far "ellipsis-h"}}</button>
